Mark GraphPersistenceWrapper.Equals(object) as obsolete.
Also, move FileGraphPersistenceWrapper to its own source file.
